producer::config takes service_name for notify
if kafka::mirror is used and  service_restart  is not set to false
then a catalog compilation error occurs.  because producer::config
will setup a notify to a non existant service.

File[/opt/kafka/config/producer.properties] { notify => Service[kafka]
}, because Service[kafka] doesn't seem to be in the catalog

so we allow producer::config to take a service_name parameter so
that when it is used by kafka::mirror::config the correct kafka-mirror
service name can be passed.