Make the systemd manifest configurable
Zookeeper sometimes does not run on the same machine as kafka, having a
hard Requires=zookeeper.service in the unit file, will make kafka not
start in these cases.

Therfore adding a parameter, to disable this Requires directive.
bastelfreak mentioned that the default could change in the next major
release.